strings and localization

+++ app/src/main/java/com/yotfr/sunmoon/presentation/task/outdated_task_list/OutdatedTaskFragment.kt
@@ -96,11 +96,13 @@ class OutdatedTaskFragment : Fragment(R.layout.fragment_outdated_task_list) {
                 return when (menuItem.itemId) {
                     R.id.mi_delete_all_tasks -> {
                         showDeleteAllDialog { deleteOption ->
-                            viewModel.onEvent(
-                                OutdatedTaskEvent.DeleteTasks(
-                                    deleteOption = deleteOption
+                            showConfirmationDialog {
+                                viewModel.onEvent(
+                                    OutdatedTaskEvent.DeleteTasks(
+                                        deleteOption = deleteOption
+                                    )
                                 )
-                            )
+                            }
                         }
                         true
                     }
@@ -248,10 +250,10 @@ class OutdatedTaskFragment : Fragment(R.layout.fragment_outdated_task_list) {
     private fun showUndoTrashSnackbar(onAction: () -> Unit) {
         Snackbar.make(
             requireView(),
-            getString(R.string.undo_delete_task_description),
+            getString(R.string.task_trashed),
             Snackbar.LENGTH_LONG
         )
-            .setAction(getString(R.string.undo_delete_task_button_text)) {
+            .setAction(getString(R.string.undo)) {
                 onAction()
             }.show()
     }
@@ -385,6 +387,19 @@ class OutdatedTaskFragment : Fragment(R.layout.fragment_outdated_task_list) {
                     0 -> onPositive(OutdatedDeleteOption.ALL_OUTDATED)
                     1 -> onPositive(OutdatedDeleteOption.ALL_OUTDATED_COMPLETED)
                 }
+            }.show()
+    }
+
+    private fun showConfirmationDialog(
+        onPositive:() -> Unit
+    ) {
+        MaterialAlertDialogBuilder(requireContext())
+            .setTitle(resources.getString(R.string.delete_tasks))
+            .setMessage(resources.getString(R.string.tasks_dialog_message))
+            .setNegativeButton(resources.getString(R.string.cancel)) { _, _ ->
+            }
+            .setPositiveButton(resources.getString(R.string.delete)) { _, _ ->
+                onPositive()
             }.show()
     }
